GeekText_Team2-zulfar/requirements.txt
alembic==0.9.9
astroid==2.2.5
autopep8==1.4.4
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
click==6.7
colorama==0.4.1
Django==2.2.2
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-MySQLdb==0.2.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
httplib2==0.11.3
idna==2.6
image==1.5.27
isort==4.3.21
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
lazy-object-proxy==1.4.1
lml==0.0.9
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.4.2.post1
numpy==1.16.4
oauth2==1.9.0.post1
oauthlib==2.0.7
Pillow==6.0.0
pycodestyle==2.5.0
pyexcel==0.5.14
pyexcel-io==0.5.18
pygeocoder==1.2.5
pylint==2.3.1
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2019.1
requests==2.22.0
requests-oauthlib==0.8.0
six==1.12.0
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
sqlparse==0.3.0
stripe==2.32.1
texttable==1.6.1
typed-ast==1.4.0
urllib3==1.22
URLObject==2.4.3
visitor==0.1.3
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.11.2
WTForms==2.1
